Bundle server specs
